Fix stack overflow crash
When queuing multiple cat images to load,
sometimes finishedLoadingAllQueuedCatImages() would
get called before all the images were queued to load,
thus causing an infinite recursion.

In this change, the caller of loadCatImageWithParameters
keeps track of how many images they have queued to load,
so it knows when to request more.